선번장관리 > 선번장목록 >  삭제 장비관리 > 장비목록 > 삭제 : 삭제하려는 장비가 선번장목록에 등록되어 있을 시 삭제할 수 없도록 로직 추가

+++ cmp_web/src/main/java/sl/qr/mh/service/eqp/eqpManageService.java
@@ -732,11 +732,19 @@ public class eqpManageService {
         returnMap.put("errorCode",false);
 
         try {
+            List<String> errorTarget = new ArrayList<>();
             for(Map<String, Object> ele : deleteList){
                 String deleteEqpTarget = ele.get("eqp_manage_id").toString();
-                eqpMapper.deleteEqpList(deleteEqpTarget);
+                int isContain = eqpMapper.checkCableIsContainEqpList(deleteEqpTarget);
+                if(isContain == 0){
+                    eqpMapper.deleteEqpList(deleteEqpTarget);
+                }
+                else{
+                    errorTarget.add(deleteEqpTarget);
+                }
             }
 
+            returnMap.put("errorTarget", errorTarget);
             returnMap.put("errorCode",true);
 
         } catch (Exception e) {
